+++ src/model/decorator.type.js
@@ -1,0 +1,25 @@
+const cast = (type, value, isNullable = false) => {
+	if (value === null && isNullable) {
+		return value;
+	}
+	return type.cast(value);
+};
+
+export default (type, isNullable = false) => {
+	return (target, property, descriptor) => {
+		if (!target.__properties__) {
+			target.__properties__ = [];
+		}
+		target.__properties__.push(property);
+		let v;
+		if (descriptor.initializer) {
+			v = cast(type, descriptor.initializer(), isNullable);
+		}
+		return {
+			get: () => v,
+			set: (value) => v = cast(type, value, isNullable),
+			enumerable: true,
+			configurable: false,
+		};
+	}
+}