+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # CHANGELOG
+
+## 3.2.0
+### Features
+- Model with type decorator.
+
 ## 3.1.0
-- updated smart-error for browser support
+### Updates
+- New version of smart-error for browser support.
 
 ## 3.0.0
-- strict validators
+### Features
+- Strict validators.
 
 ## 2.0.0
-- using default export
+### Updates
+- Using default export.